requirements.txt
# Password Strength Auditor - Dependencies
# Install with: pip install -r requirements.txt

# HTTP library for API requests
requests>=2.28.0

# Async HTTP library for concurrent API requests
aiohttp>=3.8.0

# Password strength estimation (Dropbox's zxcvbn)
zxcvbn-python>=4.4.24

# Rich text and beautiful formatting in the terminal
rich>=13.0.0

# Environment variable management
python-dotenv>=1.0.0

# YAML configuration file support
PyYAML>=6.0.0

# TOML configuration file support (Python < 3.11)
tomli>=2.0.0; python_version<"3.11"

# Data validation for configuration
pydantic>=2.0.0

# Testing dependencies (install with: pip install -r requirements.txt)
pytest>=7.0.0
pytest-cov>=4.0.0
pytest-mock>=3.10.0
pytest-asyncio>=0.21.0
